refactor: 불필요한 클래스 삭제 및 컬럼명 변경 반영

+++ src/main/kotlin/com/example/team/haribo/goms/domain/auth/service/impl/SignupServiceImpl.kt
@@ -47,7 +47,7 @@ class SignupServiceImpl(
             email = request.email,
             password = encoded,
             name = request.name,
-            grade = request.grade,
+            grade = request.grade.toInt(),
             department = request.department,
             gender = request.gender,
             role = Role.ROLE_STUDENT
